+++ src/tests/users.spec.ts
@@ -1,5 +1,4 @@
-import request from "supertest";
-import app from '../app';
+import { post, get, deleteR } from './utils';
 
 const firstUser = {
     username: 'usertest123',
@@ -102,28 +101,3 @@ describe('Delete a message', () => {
     });
 });
 
-export function post(url: string, body: object){
-    const httpRequest = request(app).post(url);
-    httpRequest.send(body);
-    httpRequest.set('Accept', 'application/json')
-    httpRequest.set('Origin', 'http://localhost:3000')
-    return httpRequest;
-}
-
-export function get(url: string, query?: object){
-    const httpRequest = request(app).get(url);
-    if (query) {
-        httpRequest.query(query);
-    }
-    httpRequest.set('Accept', 'application/json')
-    httpRequest.set('Origin', 'http://localhost:3000')
-    return httpRequest;
-}
-
-export function deleteR(url: string, body?: object){
-    const httpRequest = request(app).delete(url);
-    httpRequest.send(body);
-    httpRequest.set('Accept', 'application/json')
-    httpRequest.set('Origin', 'http://localhost:3000')
-    return httpRequest;
-}
